rename serial_test directory to data_controllers
